requirements.txt

flask
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
umap-learn
hdbscan
